ML Models/Intruder Detection/requirements.txt
paho-mqtt
face_recognition
flask
numpy
opencv-python
requests